+++ Sources/SpotifyWebAPI/Authorization/SpotifyScopeAuthorizationManager.swift
@@ -8,8 +8,6 @@ import Foundation
  conform to this protocol. `ClientCredentialsFlowManager`
  is not a conforming type because it does not support authorization scopes.
  
- You should usually not need conform additional types to this protocol.
- 
  [1]: https://developer.spotify.com/documentation/general/guides/authorization-guide/#authorization-code-flow-with-proof-key-for-code-exchange-pkce
  */
 public protocol SpotifyScopeAuthorizationManager: SpotifyAuthorizationManager { }