+++ Sources/SpotifyWebAPI/Other/LogHandler.swift
@@ -89,7 +89,6 @@ public struct SpotifyAPILogHandler: LogHandler {
 
 }
 
-
 public extension Logger {
     
     /**