+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2020-10-27
+
+### Added
+
+* Added `SpotifyAPI.authorizationManagerDidDeauthorize`. This publisher emits when the `deauthorize()` method of the authorization manager is called; `SpotifyAPI.authorizationManagerDidChange` no longer emits when `deauthorize()`  is called.
+
+### Fixed
+
+* Fixed bug in which creating an instance of `SpotifyAPI` using `init(from:)` did not properly setup the subscription to `authorizationManager.didChange`.
+
 ## [0.8.1] - 2020-10-24
 
 ### Fixed
@@ -184,6 +194,7 @@ All of the Spotify web API endpoints are now supported!
 ## [0.2.0] - 2020-09-15
 
 ### Changed
+
 - Refactored SpotifyAPI methods that require either authorization scopes or an access token that was retrieved for a user into conditional extensions where  AuthorizationManager conforms SpotifyScopeAuthorizationManager. This new protocol extends SpotifyAuthorizationManager and requires that conforming types support authorization scopes. Currently, only `AuthorizationCodeFlowManager` conforms to this protocol, but a future version of this library will support the [Authorization Code Flow with Proof Key for Code Exchange](https://developer.spotify.com/documentation/general/guides/authorization-guide/#authorization-code-flow-with-proof-key-for-code-exchange-pkce), which will also conform to this protocol. `ClientCredentialsFlowManager` is not a conforming type because it does not support authorization scopes. This change provides a compile-time guarantee that you can not call methods that require authorization scopes when using the `ClientCredentialsFlowManager`.
 - Refactored `resumePlayback` into two separate methods: `resumePlayback(deviceId:)` only resumes the user's current playback. `play(_:deviceId:)` (added) plays specific content for the current user.
 - When multiple asyncronous requests are made to refresh the access token, only one network request will be made. While this request is in progress, additional requests to refresh the access token will receive the same publisher as a class instance.